refactor(app): extract PlayerPane; PlayerScreen becomes thin host (app-21)
